consolidate Rust outstation examples
